example/requirements.txt
aiohttp-login
jinja_app_loader
asyncpg
markdown
